# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vendas API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
        "/orders": {
            "post": {
                "tags": [
                    "/order"
                ],
                "summary": "Criar Venda",
                "operationId": "445bab4f7dc9ea6fcd54f286c0209478",
                "parameters": [
                    {
                        "name": "seller_id",
                        "in": "query",
                        "style": "form",
                        "schema": {
                            "type": "bigint"
                        }
                    },
                    {
                        "name": "value",
                        "in": "query",
                        "style": "form",
                        "schema": {
                            "type": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            }
        },
        "/orders/total/{date}": {
            "get": {
                "tags": [
                    "/order"
                ],
                "summary": "Mostrar o total de vendas do dia",
                "operationId": "91de3d44c6d1bb939c69523fbf2d739b",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        },
                        "example": "13-06-2022"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            }
        },
        "/sellers": {
            "get": {
                "tags": [
                    "/seller"
                ],
                "summary": "Listar todos os vendedores",
                "operationId": "47baf47a7adcb6cd7d51464a887e7074",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "/seller"
                ],
                "summary": "Criar vendedor",
                "operationId": "01819d3682e43f1ef261c8210301ac7c",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            }
        },
        "/sellers/{seller}/orders": {
            "get": {
                "tags": [
                    "/seller"
                ],
                "summary": "Mostrar todas as vendas do vendedor",
                "operationId": "e9e4ca1bc19fa2b6c669f0bbf1198369",
                "parameters": [
                    {
                        "name": "seller",
                        "in": "path",
                        "style": "form",
                        "schema": {
                            "type": "bigint"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            }
        }
    }
}